fix mangeUI to modifyUI

+++ src/ManageUI.java
@@ -68,7 +68,7 @@ public class ManageUI extends JPanel {
 				int sn = (int) orderlistTable.getModel().getValueAt(orderlistTable.getSelectedRow(), 9); 
 				int dn = (int) orderlistTable.getModel().getValueAt(orderlistTable.getSelectedRow(), 10); 
 				int qn = (int) orderlistTable.getModel().getValueAt(orderlistTable.getSelectedRow(), 11);
-				int sumPrice = (int) orderlistTable.getModel().getValueAt(orderlistTable.getSelectedRow(), 12);
+				long sumPrice = (long) orderlistTable.getModel().getValueAt(orderlistTable.getSelectedRow(), 12);
 				mUIMainFrame.changeUI(UIMainFrame.UIStage.MANAGE,UIMainFrame.UIStage.MODIFY,reservationID,CID,COD,HotelID,sn,dn,qn,sumPrice);
 				revise.setForeground(Color.black);
 			}
